requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.1.1
Werkzeug==3.1.3
